functions/requirements.txt
functions-framework==3.4.0
google-generativeai==0.3.2
flask==2.3.3
